Python apps fix

See merge request !35

Former-commit-id: 9c7b295cb9c900e5fef21e59969131887461e66a

+++ python/examples/qr/src/qr.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 #
-#  Copyright 2002-2018 Barcelona Supercomputing Center (www.bsc.es)
+#  Copyright 2002-2019 Barcelona Supercomputing Center (www.bsc.es)
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -227,7 +227,7 @@ def joinMatrix(A):
 
 def splitMatrix(A, MSIZE):
     splittedMatrix = []
-    bSize = len(A) / MSIZE
+    bSize = int(len(A) / MSIZE)
     for i in range(MSIZE):
         splittedMatrix.append([])
         for j in range(MSIZE):
